sentry:  Move SentryTracer to sentry.py
